Fix breakage when using a custom cookie jar
As stated by the documentation of requests:

Requests does not use the dict interface internally; it's just for
compatibility with external client code. All requests code should work
out of the box with externally provided instances of ``CookieJar``, e.g.
``LWPCookieJar`` and ``FileCookieJar``

Using RequestsCookieJar's get method breaks when the user wishes to store
a custom instance of cookie jar in the session, such as LWPCookieJar.

The _find_no_duplicates function is based on RequestsCookieJar's internal method.

This change was tested in both Python 3 and 2