Update test_adapters.py
I made some modification to the previous version of the code . In this version of the script, I used pytest as the test framework and added a fixture to create and close the CloudflareAdapter instance automatically. I also replaced the sure library with the built-in assert statement and added more tests for the CloudflareAdapter class.